requirements.txt
flask==3.0.3
python-telegram-bot==21.6
google-generativeai==0.7.2
google-api-python-client==2.139.0
google-auth==2.32.0
gunicorn==22.0.0
uvicorn==0.30.6
